+++ pymc3/plots/posteriorplot.py
@@ -6,7 +6,7 @@ from .artists import plot_posterior_op
 from .utils import identity_transform, get_default_varnames
 
 
-def plot_posterior(trace, varnames=None, transform=identity_transform, figsize=None,
+def plot_posterior(trace, varnames=None, transform=identity_transform, figsize=None, text_size=16,
                    alpha_level=0.05, round_to=3, point_estimate='mean', rope=None,
                    ref_val=None, kde_plot=False, plot_transformed=False, ax=None, **kwargs):
     """Plot Posterior densities in style of John K. Kruschke book.
@@ -21,6 +21,8 @@ def plot_posterior(trace, varnames=None, transform=identity_transform, figsize=N
         Function to transform data (defaults to identity)
     figsize : figure size tuple
         If None, size is (12, num of variables * 2) inch
+    text_size : int
+        Text size of the point_estimates and HPD (Default:16)
     alpha_level : float
         Defines range for High Posterior Density
     round_to : int
@@ -43,7 +45,6 @@ def plot_posterior(trace, varnames=None, transform=identity_transform, figsize=N
         value of the argument kde_plot
         Some defaults are added, if not specified
         color='#87ceeb' will match the style in the book
-
 
     Returns
     -------
@@ -79,9 +80,9 @@ def plot_posterior(trace, varnames=None, transform=identity_transform, figsize=N
             figsize = (6, 2)
         if ax is None:
             fig, ax = plt.subplots()
-        plot_posterior_op(transform(trace), ax=ax, kde_plot=kde_plot,
+        plot_posterior_op(transform(trace), figsize=figsize, ax=ax, kde_plot=kde_plot,
                           point_estimate=point_estimate, round_to=round_to,
-                          alpha_level=alpha_level, ref_val=ref_val, rope=rope, **kwargs)
+                          alpha_level=alpha_level, ref_val=ref_val, rope=rope, text_size=text_size, **kwargs)
     else:
         if varnames is None:
             varnames = get_default_varnames(trace.varnames, plot_transformed)
@@ -93,13 +94,14 @@ def plot_posterior(trace, varnames=None, transform=identity_transform, figsize=N
 
         for a, v in zip(np.atleast_1d(ax), trace_dict):
             tr_values = transform(trace_dict[v])
-            plot_posterior_op(tr_values, ax=a, kde_plot=kde_plot,
+            plot_posterior_op(tr_values, figsize=figsize, ax=a, kde_plot=kde_plot,
                               point_estimate=point_estimate, round_to=round_to,
-                              alpha_level=alpha_level, ref_val=ref_val, rope=rope, **kwargs)
+                              alpha_level=alpha_level, ref_val=ref_val, rope=rope, text_size=text_size, **kwargs)
             a.set_title(v)
 
         plt.tight_layout()
     return ax
+
 
 def plot_posterior_predictive_glm(trace, eval=None, lm=None, samples=30, **kwargs):
     """Plot posterior predictive of a linear model.